+++ app/src/main/java/org/simple/clinic/facility/alertchange/AlertFacilityChangeEffect.kt
@@ -1,3 +1,11 @@
 package org.simple.clinic.facility.alertchange
 
-sealed class AlertFacilityChangeEffect
+sealed interface AlertFacilityChangeEffect {
+  data object LoadIsFacilityChangedStatus : AlertFacilityChangeEffect
+
+  data object MarkFacilityChangedAsFalse : AlertFacilityChangeEffect
+}
+
+sealed interface AlertFacilityChangeViewEffect : AlertFacilityChangeEffect {
+  data object CloseSheetWithContinuation : AlertFacilityChangeViewEffect
+}