+++ app/src/main/java/org/simple/clinic/facility/alertchange/AlertFacilityChangeUi.kt
@@ -1,0 +1,6 @@
+package org.simple.clinic.facility.alertchange
+
+interface AlertFacilityChangeUi {
+  fun showFacilityChangeAlert()
+  fun hideFacilityChangeAlert()
+}