+++ app/src/main/java/org/simple/clinic/facility/alertchange/AlertFacilityChangeUiRenderer.kt
@@ -1,0 +1,16 @@
+package org.simple.clinic.facility.alertchange
+
+import org.simple.clinic.mobius.ViewRenderer
+
+class AlertFacilityChangeUiRenderer(
+    private val ui: AlertFacilityChangeUi
+) : ViewRenderer<AlertFacilityChangeModel> {
+
+  override fun render(model: AlertFacilityChangeModel) {
+    if (model.isFacilityChanged) {
+      ui.showFacilityChangeAlert()
+    } else {
+      ui.hideFacilityChangeAlert()
+    }
+  }
+}