+++ app/src/main/java/org/simple/clinic/facility/alertchange/UiActions.kt
@@ -1,0 +1,5 @@
+package org.simple.clinic.facility.alertchange
+
+interface UiActions {
+  fun closeSheetWithContinuation()
+}